NBSidePanel migrated to TreeView. Desktop files now show Icons. Over-sized are icons now scaled to the correct size.

+++ Gui/Others/Settings/NBSettings.hpp
@@ -26,7 +26,7 @@ class NBSettings {
 				QString IconTheme;
 				bool ImagePreviews;
 				bool NativeTitleBar;
-				quint64 MaxIOJobs;
+				int MaxIOJobs;
 		} General;
 
 		// Session Settings
